refactor: compile applets as multicall binary
