+++ .gitignore
@@ -1,5 +1,5 @@
 # Cargo
-.cargo.
+.cargo
 target/
 build/
 _build/
