+++ cosmic-applet-audio/Cargo.toml
@@ -11,7 +11,7 @@ libpulse-glib-binding = "2.28.1"
 tokio = { version = "1.20.1", features=["full"] }
 libcosmic.workspace = true
 cosmic-time.workspace = true
-tracing = "0.1.40"
+tracing.workspace = true
 tracing-subscriber.workspace = true
 tracing-log.workspace = true
 # Application i18n
