+++ cosmic-applet-battery/src/config.rs
@@ -1,3 +1,1 @@
 pub const APP_ID: &str = "com.system76.CosmicAppletButton";
-pub const PROFILE: &str = "";
-pub const VERSION: &str = "0.1.0";
