+++ cosmic-applet-minimize/Cargo.toml
@@ -13,7 +13,7 @@ memmap2 = "0.9.0"
 rustix = { version = "0.38.0", features = ["fs"] }
 png = "0.17.5"
 tokio = { version = "1.17.0", features = ["sync", "macros"] }
-tracing = "0.1.40"
+tracing.workspace = true
 tracing-subscriber.workspace = true
 tracing-log.workspace = true
 tempfile = "3.5.0"
