+++ cosmic-applet-network/Cargo.toml
@@ -13,12 +13,13 @@ cosmic-time.workspace = true
 futures = "0.3"
 zbus.workspace = true
 log = "0.4"
-tracing-subscriber.workspace = true
+tracing.workspace = true
 tracing-log.workspace = true
+tracing-subscriber.workspace = true
 itertools = "0.10.3"
 slotmap = "1.0.6"
 tokio = { version = "1.15.0", features = ["full"] }
-anyhow.workspace = true 
+anyhow.workspace = true
 # Application i18n
 i18n-embed = { version = "0.13.4", features = ["fluent-system", "desktop-requester"] }
 i18n-embed-fl = "0.6.4"
