+++ cosmic-applet-notifications/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 license = "GPL-3.0-or-later"
 
 [dependencies]
-anyhow.workspace = true 
+anyhow.workspace = true
 libcosmic.workspace = true
 cosmic-time.workspace = true
 nix = "0.26"
@@ -22,12 +22,12 @@ cosmic-notifications-util = { git = "https://github.com/pop-os/cosmic-notificati
 cosmic-notifications-config = { git = "https://github.com/pop-os/cosmic-notifications" }
 # cosmic-notifications-util = { path = "../../cosmic-notifications-daemon/cosmic-notifications-util" }
 # cosmic-notifications-config = { path = "../../cosmic-notifications-daemon/cosmic-notifications-config" }
-tracing = "0.1"
+tracing.workspace = true
+tracing-subscriber.workspace = true
+tracing-log.workspace = true
 ron = "0.8"
 sendfd = { version = "0.4", features = ["tokio"] }
 bytemuck = "1"
-tracing-subscriber.workspace = true
-tracing-log.workspace = true
 zbus.workspace = true
 # Application i18n
 i18n-embed = { version = "0.13.4", features = [
