+++ cosmic-applet-power/Cargo.toml
@@ -20,4 +20,6 @@ i18n-embed-fl = "0.6"
 rust-embed = "6.6"
 rust-embed-utils = "7.5.0"
 once_cell = "1.17.1"
-
+tracing.workspace = true
+tracing-subscriber.workspace = true
+tracing-log.workspace = true
