+++ cosmic-applet-status-area/Cargo.toml
@@ -9,4 +9,7 @@ futures = "0.3"
 libcosmic.workspace = true
 serde = "1"
 tokio = { version = "1.23.0" }
+tracing.workspace = true
+tracing-log.workspace = true
+tracing-subscriber.workspace = true
 zbus.workspace = true
