+++ cosmic-applet-status-area/src/lib.rs
@@ -1,0 +1,6 @@
+mod components;
+mod subscriptions;
+
+pub fn run() -> cosmic::iced::Result {
+    components::app::main()
+}
