+++ cosmic-applet-status-area/src/main.rs
@@ -1,6 +1,10 @@
-mod components;
-mod subscriptions;
+const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 fn main() -> cosmic::iced::Result {
-    components::app::main()
+    tracing_subscriber::fmt::init();
+    let _ = tracing_log::LogTracer::init();
+
+    tracing::info!("Starting status-area applet with version {VERSION}");
+
+    cosmic_applet_status_area::run()
 }
