+++ cosmic-applet-tiling/Cargo.toml
@@ -24,6 +24,8 @@ once_cell = "1"
 i18n-embed = { version = "0.14.0", features = ["fluent-system", "desktop-requester"] }
 i18n-embed-fl = "0.7.0"
 rust-embed = "8.0.0"
-tracing = "0.1"
+tracing.workspace = true
+tracing-subscriber.workspace = true
+tracing-log.workspace = true
 cosmic-comp-config = { git = "https://github.com/pop-os/cosmic-comp.git", rev = "5eb5af4" }
 tokio = { version = "1.17.0", features = ["sync", "rt"] }
