+++ cosmic-applet-time/Cargo.toml
@@ -12,6 +12,8 @@ chrono = { version = "0.4.34", features = ["clock"] }
 once_cell = "1"
 tokio = { version = "1.36.0", features = ["time"] }
 tracing.workspace = true
+tracing-log.workspace = true
+tracing-subscriber.workspace = true
 
 # Application i18n
 i18n-embed = { version = "0.13.4", features = ["fluent-system", "desktop-requester"] }
