+++ cosmic-applet-time/src/lib.rs
@@ -1,0 +1,11 @@
+mod localize;
+mod time;
+mod window;
+
+use window::Window;
+
+pub fn run() -> cosmic::iced::Result {
+    localize::localize();
+
+    cosmic::applet::run::<Window>(true, ())
+}
