+++ cosmic-applet-workspaces/Cargo.toml
@@ -9,13 +9,13 @@ libcosmic.workspace = true
 cctk.workspace = true
 cosmic-protocols.workspace = true
 nix = "0.27.1"
-tracing = "0.1"
+tracing.workspace = true
 tracing-subscriber.workspace = true
 tracing-log.workspace = true
 once_cell = "1.9"
 futures = "0.3.21"
 xdg = "2.4.0"
-anyhow.workspace = true 
+anyhow.workspace = true
 tokio = "1.35"
 # Application i18n
 i18n-embed = { version = "0.13.4", features = ["fluent-system", "desktop-requester"] }
