+++ cosmic-applet-workspaces/src/lib.rs
@@ -1,0 +1,16 @@
+mod components;
+#[rustfmt::skip]
+mod config;
+mod localize;
+mod wayland;
+mod wayland_subscription;
+
+use localize::localize;
+
+use crate::components::app;
+
+pub fn run() -> cosmic::iced::Result {
+    localize();
+
+    app::run()
+}
