+++ cosmic-applet-workspaces/src/main.rs
@@ -1,28 +1,10 @@
-mod components;
-#[rustfmt::skip]
-mod config;
-mod localize;
-mod wayland;
-mod wayland_subscription;
-
-use tracing::info;
-
-use localize::localize;
-
-use crate::components::app;
-
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 fn main() -> cosmic::iced::Result {
-    // Initialize logger
     tracing_subscriber::fmt::init();
     let _ = tracing_log::LogTracer::init();
 
-    info!("Starting audio applet with version {VERSION}");
-    info!("Iced Workspaces Applet ({VERSION})");
+    tracing::info!("Starting workspaces applet with version {VERSION}");
 
-    // Prepare i18n
-    localize();
-
-    app::run()
+    cosmic_applet_workspaces::run()
 }
