+++ debian/rules
@@ -16,15 +16,15 @@ override_dh_auto_clean:
 	fi
 
 	if ! ischroot && test "${VENDOR}" = "1"; then \
-		mkdir -p .cargo; \
-		cargo vendor --sync Cargo.toml | head -n -1 > .cargo/config; \
-		echo 'directory = "vendor"' >> .cargo/config; \
-		tar pcf vendor.tar vendor; \
-		rm -rf vendor; \
+		just vendor; \
 	fi
 
 override_dh_auto_build:
-	just rootdir=$(DESTDIR) debug=$(DEBUG) vendor=$(VENDOR) build
+	if test "${VENDOR}" = "1"; then \
+		just rootdir=$(DESTDIR) debug=$(DEBUG) build-vendored; \
+	else \
+		just rootdir=$(DESTDIR) debug=$(DEBUG) build-release; \
+	fi
 
 override_dh_auto_install:
 	just rootdir=$(DESTDIR) install
